rename elementType and remove duplicate (#371)
