+++ src/components/DirectoryItemSelector/directory-item-selector.tsx
@@ -12,7 +12,8 @@ import {
     useRef,
     useState,
 } from 'react';
-import { getFileIcon, elementType } from '../../utils/ElementType.jsx';
+import { getFileIcon } from '../../utils/ElementIcon';
+import { ElementType } from '../../utils/ElementType';
 import { useSnackMessage } from '../../hooks/useSnackMessage.js';
 import TreeViewFinder from '../TreeViewFinder';
 import { Theme } from '@mui/material';
@@ -86,7 +87,7 @@ const DirectoryItemSelector: FunctionComponent<DirectoryItemSelectorProps> = ({
     const openRef = useRef<boolean>();
     openRef.current = open;
     const contentFilter = useCallback(
-        () => new Set([elementType.DIRECTORY, ...types]),
+        () => new Set([ElementType.DIRECTORY, ...types]),
         [types]
     );
 
@@ -98,11 +99,11 @@ const DirectoryItemSelector: FunctionComponent<DirectoryItemSelectorProps> = ({
                 specificMetadata: e.specificMetadata,
                 icon: getFileIcon(e.type, styles.icon),
                 children:
-                    e.type === elementType.DIRECTORY
+                    e.type === ElementType.DIRECTORY
                         ? convertChildren(e.children)
                         : undefined,
                 childrenCount:
-                    e.type === elementType.DIRECTORY
+                    e.type === ElementType.DIRECTORY
                         ? e.subdirectoriesCount
                         : undefined,
             };
@@ -117,13 +118,13 @@ const DirectoryItemSelector: FunctionComponent<DirectoryItemSelectorProps> = ({
                     name: e.elementName,
                     icon: getFileIcon(e.type, styles.icon),
                     children:
-                        e.type === elementType.DIRECTORY
+                        e.type === ElementType.DIRECTORY
                             ? convertChildren(
                                   nodeMap.current[e.elementUuid].children
                               )
                             : undefined,
                     childrenCount:
-                        e.type === elementType.DIRECTORY
+                        e.type === ElementType.DIRECTORY
                             ? e.subdirectoriesCount
                             : undefined,
                 };
@@ -195,7 +196,7 @@ const DirectoryItemSelector: FunctionComponent<DirectoryItemSelectorProps> = ({
                             const children = itemFilter
                                 ? childrenWithMetadata.filter((val: any) => {
                                       // Accept every directory
-                                      if (val.type === elementType.DIRECTORY) {
+                                      if (val.type === ElementType.DIRECTORY) {
                                           return true;
                                       }
                                       // otherwise filter with the custom itemFilter func
