+++ src/utils/ElementIcon.jsx
@@ -13,38 +13,26 @@ import {
     Settings as SettingsIcon,
     NoteAlt as NoteAltIcon,
 } from '@mui/icons-material';
-
-export const elementType = {
-    DIRECTORY: 'DIRECTORY',
-    STUDY: 'STUDY',
-    CASE: 'CASE',
-    FILTER: 'FILTER',
-    MODIFICATION: 'MODIFICATION',
-    CONTINGENCY_LIST: 'CONTINGENCY_LIST',
-    VOLTAGE_INIT_PARAMETERS: 'VOLTAGE_INIT_PARAMETERS',
-    SECURITY_ANALYSIS_PARAMETERS: 'SECURITY_ANALYSIS_PARAMETERS',
-    LOADFLOW_PARAMETERS: 'LOADFLOW_PARAMETERS',
-    SENSITIVITY_PARAMETERS: 'SENSITIVITY_PARAMETERS',
-};
+import { ElementType } from './ElementType';
 
 export function getFileIcon(type, style) {
     switch (type) {
-        case elementType.STUDY:
+        case ElementType.STUDY:
             return <PhotoLibraryIcon sx={style} />;
-        case elementType.CASE:
+        case ElementType.CASE:
             return <PhotoIcon sx={style} />;
-        case elementType.CONTINGENCY_LIST:
+        case ElementType.CONTINGENCY_LIST:
             return <OfflineBoltIcon sx={style} />;
-        case elementType.MODIFICATION:
+        case ElementType.MODIFICATION:
             return <NoteAltIcon sx={style} />;
-        case elementType.FILTER:
+        case ElementType.FILTER:
             return <ArticleIcon sx={style} />;
-        case elementType.VOLTAGE_INIT_PARAMETERS:
-        case elementType.SECURITY_ANALYSIS_PARAMETERS:
-        case elementType.LOADFLOW_PARAMETERS:
-        case elementType.SENSITIVITY_PARAMETERS:
+        case ElementType.VOLTAGE_INIT_PARAMETERS:
+        case ElementType.SECURITY_ANALYSIS_PARAMETERS:
+        case ElementType.LOADFLOW_PARAMETERS:
+        case ElementType.SENSITIVITY_PARAMETERS:
             return <SettingsIcon sx={style} />;
-        case elementType.DIRECTORY:
+        case ElementType.DIRECTORY:
             // to easily use in TreeView we do not give icons for directories
             return;
         default:
